Separate standard image nodes
