+++ backend/src/nodes/nodes/image/iterator/image_file_iterator.py
@@ -8,7 +8,7 @@ from process import IteratorContext
 from sanic.log import logger
 
 from ....categories import ImageCategory
-from ..image_nodes import ImReadNode
+from ..load_image import ImReadNode
 from ....node_base import IteratorNodeBase, NodeBase
 from ....node_factory import NodeFactory
 from ....properties.inputs import IteratorInput, DirectoryInput
